+++ packages/expo-screen-orientation/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 💡 Others
 
 - [plugin] Added SDK 43 tests for new AppDelegate template ([#14763](https://github.com/expo/expo/pull/14763) by [@EvanBacon](https://github.com/EvanBacon))
+- The app delegate subcontractor on iOS has been separated from the singleton module to hook into the new implementation of `AppDelegateWrapper`. ([#14867](https://github.com/expo/expo/pull/14867) by [@tsapeta](https://github.com/tsapeta))
 
 ## 4.0.1 — 2021-10-01